Add workflow to build container image on commit
This will automatically build a container image and push it into ghcr.io on
every commit. The image will be tagged with the branch name, the commit
SHA, and the commit tag, if this push corresponds to a tag.

The `latest` tag will correspond to the most recently pushed git tag.